Move shader-compiler to its own file
